Move authHook to utils/hooks folder
